requirements.txt
discord.py
flask
python-dotenv
better-profanity
joblib